refactor: create separate custom hook to manage state of share-playground-modal & break the component into smaller parts
